+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Add option to hide default sorting button.
 * `flexberry-groupedit` component:
     * Add check all at page button.
+* `flexberry-lookup` component:
+    * Add property `autofillByLimit`.
 
 ### Changed
 * `flexberry-lookup` component: